feat : cityRepository 추가, 관리비 합쳐서 total_price 저장

+++ src/main/java/sion/bestRoom/dto/DabangRoomDTO.java
@@ -15,7 +15,7 @@ public class DabangRoomDTO {
     private String title;
     private String room_desc2; // 방 설명 // 층 , 면적 , 관리비가 있음 현재 ex ) 16.52m², 관리비 7만"
     private String img_url;
-    private Long selling_type; // 0:월세 , 1: 전세  2: 매매
+    private Integer selling_type; // 0:월세 , 1: 전세  2: 매매
     private String selling_type_str;
 
     private List<Double> location;
